Update README.md
This repository stands as supplementary materials for the publication "New rolling horizon optimization approaches to balance short-term and long-term
decisions: an application to energy planning".

The published scripts are use to compute the cost functions as defined in the publication. They can not be used directly since the computations ultimately rely on the in-house Persee software from the LSET laboratory. Still, authors hope this can help to clarify the methodology if the need occurs.